requirements.txt
cryptography==46.0.3
fastapi==0.128.0
filelock==3.20.1
firecrawl_py==4.9.0
Jinja2==3.0.3
langchain==1.2.0
langchain_classic==1.0.1
langchain_core==1.2.5
langchain_openai==1.1.6
langgraph==1.0.5
numpy==2.4.0
protobuf==6.33.2
pydantic==2.12.5
pyOpenSSL==25.3.0
python-dotenv==1.2.1
qdrant_client==1.16.2
redis==7.1.0
setuptools==74.1.2
shopifyapi==12.7.0
structlog==25.5.0
tornado==6.4.2
urllib3_secure_extra==0.1.0
uvicorn==0.40.0
pytest-asyncio>=0.21.0
